Worked on layout system, not yet successful

+++ src/node_data.rs
@@ -9,17 +9,17 @@ pub enum FlexDirection {
 
 #[derive(Debug)]
 pub struct NodeData {
-    pub min_width_rem: Option<u32>,
-    pub min_height_rem: Option<u32>,
-    pub max_width_rem: Option<u32>,
-    pub max_height_rem: Option<u32>,
+    pub min_width_rem: Option<f32>,
+    pub min_height_rem: Option<f32>,
+    pub max_width_rem: Option<f32>,
+    pub max_height_rem: Option<f32>,
 
     /// Width of the node. 
     /// **WARNING**: Must be initialized for the root node
-    pub width: Option<u32>,
+    pub width: Option<f32>,
     /// Width of the node. 
     /// **WARNING**: Must be initialized for the root node
-    pub height: Option<u32>,
+    pub height: Option<f32>,
     /// What direction the children should flex to
     pub flex_direction: FlexDirection,
     /// **DEBUG** color of the node
@@ -28,12 +28,12 @@ pub struct NodeData {
 
 impl NodeData {
     /// Creates a new node
-    pub fn new(min_width_rem: Option<u32>, 
-               min_height_rem: Option<u32>, 
-               max_width_rem: Option<u32>, 
-               max_height_rem: Option<u32>,
-               width: Option<u32>,
-               height: Option<u32>,
+    pub fn new(min_width_rem: Option<f32>, 
+               min_height_rem: Option<f32>, 
+               max_width_rem: Option<f32>, 
+               max_height_rem: Option<f32>,
+               width: Option<f32>,
+               height: Option<f32>,
                flex_direction: FlexDirection,
                debug_color: DebugColor)
     -> Self {
